refactor: properly rename files
